feat(dbpedia-lookup): refactor controller to use a service
A controller should only parse parameters and call service methods in order to separate concerns.
